% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizar_function.R
\name{organizaR}
\alias{organizaR}
\title{Create a folder directory displayed as a message}
\usage{
organizaR()
}
\value{
Folder structure
}
\description{
Create a folder directory displayed as a message
}
